% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{coef.stoch_logistic}
\alias{coef.stoch_logistic}
\title{Retrieve fitted coefficients from stochastic logistic regression object}
\usage{
\method{coef}{stoch_logistic}(object, ...)
}
\arguments{
\item{object}{A `stoch_logistic` object as output by function `stochastic.logistic.regression`.
Must have already been fit to at least 1 batch of data.}

\item{...}{Not used.}
}
\value{
An (n x 1) matrix with the coefficients, in the same format as those from `glm`.
}
\description{
Retrieve fitted coefficients from stochastic logistic regression object
}
\seealso{
\link{stochastic.logistic.regression}
}
