% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers_guided.R
\name{get_curr_x}
\alias{get_curr_x}
\title{Get current values of the optimization variables}
\usage{
get_curr_x(optimizer)
}
\arguments{
\item{optimizer}{An optimizer (guided-mode) from this module, as output by
functions `oLBFGS`, `SQN`, `adaQN`.}
}
\value{
A numeric vector with the current values of the variables being optimized.
}
\description{
Get current values of the optimization variables
}
\seealso{
\link{oLBFGS} , \link{SQN} , \link{adaQN}
}
