% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers_guided.R
\name{get_iteration_number}
\alias{get_iteration_number}
\title{Get current iteration number from the optimizer object}
\usage{
get_iteration_number(optimizer)
}
\arguments{
\item{optimizer}{An optimizer (guided-mode) from this module, as output by
functions `oLBFGS`, `SQN`, `adaQN`.}
}
\value{
The current iteration number.
}
\description{
Get current iteration number from the optimizer object
}
\seealso{
\link{oLBFGS} , \link{SQN} , \link{adaQN}
}
