% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers_guided.R
\name{partial_fit}
\alias{partial_fit}
\title{Partial fit stochastic model to new data}
\usage{
partial_fit(optimizer, X, y = NULL, weights = NULL, ...)
}
\arguments{
\item{optimizer}{A stochastic optimizer from this package as output by functions `oLBFGS`, `SQN`, `adaQN`.
Will be modified in-place.}

\item{X}{Covariates to pass to the user-defined gradient / objective / Hessian-vector functions.}

\item{y}{Target variable to pass to the user-defined gradient / objective / Hessian-vector functions.}

\item{weights}{Target variable to pass to the user-defined gradient / objective / Hessian-vector functions.}

\item{...}{Additional arguments to pass to the user-defined gradient / objective / Hessian-vector functions.}
}
\value{
No return value (object is modified in-place).
}
\description{
Runs one iteration of the stochastic optimizer on the new data passed here.
}
\seealso{
\link{oLBFGS} , \link{SQN} , \link{adaQN}
}
