% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{predict.stoch_logistic}
\alias{predict.stoch_logistic}
\title{Prediction function for stochastic logistic regression}
\usage{
\method{predict}{stoch_logistic}(object, newdata, type = "prob", ...)
}
\arguments{
\item{object}{A `stoch_logistic` object as output by function `stochastic.logistic.regression`.}

\item{newdata}{New data on which to make predictions.}

\item{type}{Type of prediction to make. Can pass `prob` to get probabilities for the positive class,
or `class` to get the predicted class.}

\item{...}{Not used.}
}
\value{
A vector with the predicted classes or probabilities for `newdata`.
}
\description{
Makes predictions for new data from the fitted model. Model have already
been fit to at least 1 batch of data.
}
\seealso{
\link{stochastic.logistic.regression}
}
