% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers_guided.R
\name{print.adaQN}
\alias{print.adaQN}
\title{Print summary info about adaQN guided-mode object}
\usage{
\method{print}{adaQN}(x, ...)
}
\arguments{
\item{x}{An `adaQN` object as output by function of the same name.}

\item{...}{Not used.}
}
\description{
Print summary info about adaQN guided-mode object
}
