% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers_free.R
\name{print.adaQN_free}
\alias{print.adaQN_free}
\title{Print summary info about adaQN free-mode object}
\usage{
\method{print}{adaQN_free}(x, ...)
}
\arguments{
\item{x}{An `adaQN_free` object as output by function of the same name.}

\item{...}{Not used.}
}
\description{
Print summary info about adaQN free-mode object
}
