% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers_guided.R
\name{print.oLBFGS}
\alias{print.oLBFGS}
\title{Print summary info about oLBFGS guided-mode object}
\usage{
\method{print}{oLBFGS}(x, ...)
}
\arguments{
\item{x}{An `oLBFGS` object as output by function of the same name.}

\item{...}{Not used.}
}
\description{
Print summary info about oLBFGS guided-mode object
}
