% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers_free.R
\name{print.oLBFGS_free}
\alias{print.oLBFGS_free}
\title{Print summary info about oLBFGS free-mode object}
\usage{
\method{print}{oLBFGS_free}(x, ...)
}
\arguments{
\item{x}{An `oLBFGS_free` object as output by function of the same name.}

\item{...}{Not used.}
}
\description{
Print summary info about oLBFGS free-mode object
}
