% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers_guided.R
\name{print.SQN}
\alias{print.SQN}
\title{Print summary info about SQN guided-mode object}
\usage{
\method{print}{SQN}(x, ...)
}
\arguments{
\item{x}{An `SQN` object as output by function of the same name.}

\item{...}{Not used.}
}
\description{
Print summary info about SQN guided-mode object
}
