% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers_free.R
\name{print.SQN_free}
\alias{print.SQN_free}
\title{Print summary info about SQN free-mode object}
\usage{
\method{print}{SQN_free}(x, ...)
}
\arguments{
\item{x}{An `SQN_free` object as output by function of the same name.}

\item{...}{Not used.}
}
\description{
Print summary info about SQN free-mode object
}
