% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{print.stoch_logistic}
\alias{print.stoch_logistic}
\title{Print general info about stochastic logistic regression object}
\usage{
\method{print}{stoch_logistic}(x, ...)
}
\arguments{
\item{x}{A `stoch_logistic` object as output by function `stochastic.logistic.regression`.}

\item{...}{Not used.}
}
\description{
Print general info about stochastic logistic regression object
}
\seealso{
\link{stochastic.logistic.regression}
}
