% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{summary.stoch_logistic}
\alias{summary.stoch_logistic}
\title{Print general info about stochastic logistic regression object}
\usage{
\method{summary}{stoch_logistic}(object, ...)
}
\arguments{
\item{object}{A `stoch_logistic` object as output by function `stochastic.logistic.regression`.}

\item{...}{Not used.}
}
\description{
Same as `print` function. To check the fitted coefficients use function `coef`.
}
\seealso{
\link{coef.stoch_logistic} , \link{stochastic.logistic.regression}
}
