% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers_free.R
\name{update_fun}
\alias{update_fun}
\title{Update objective function value (adaQN)}
\usage{
update_fun(optimizer, fun)
}
\arguments{
\item{optimizer}{An `adaQN_free` object which after the last run had requested a new function evaluation.}

\item{fun}{Function as evaluated (in expectation) on the values of `x` that were returned in the request.}
}
\value{
No return value (object is updated in-place).
}
\description{
Update the (expected) value of the objective function in an `adaQN_free` object, after
it has been requested by the optimizer (do NOT update it otherwise).
}
