% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers_free.R
\name{update_gradient}
\alias{update_gradient}
\title{Update gradient (oLBFGS, SQN, adaQN)}
\usage{
update_gradient(optimizer, gradient)
}
\arguments{
\item{optimizer}{A free-mode optimizer from this package (`oLBFGS_free`, `SQN_free`, `adaQN_free`) which
after the last run had requested a new gradient evaluation..}

\item{gradient}{The (expected value of the) gradient as evaluated on the values of `x` that
were returned in the request. Must be a numeric vector.}
}
\value{
No return value (object is updated in-place).
}
\description{
Update the (expected) gradient in an optimizer from this package, after
it has been requested by the optimizer (do NOT update it otherwise).
}
